requirements.txt
aiohttp==3.9.1
asyncpg==0.29.0
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
contextlib2==21.6.0
datadog-api-client==2.31.0
idna==3.4
nats-py==2.2.0
psycopg2-binary==2.9.6
pydantic==2.5.3
pydantic-settings==2.1.0
pycodestyle==2.10.0
pycountry==24.6.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
requests==2.28.2
schema==0.7.5
six==1.17.0
tenacity==8.2.3
tomli==2.0.1
typing_extensions==4.12.2
